Move internal UI.setSession to UIInternals
