+++ hummingbird-server/src/main/java/com/vaadin/ui/UI.java
@@ -48,7 +48,6 @@ import com.vaadin.server.VaadinService;
 import com.vaadin.server.VaadinServlet;
 import com.vaadin.server.VaadinSession;
 import com.vaadin.server.communication.PushConnection;
-import com.vaadin.shared.communication.PushMode;
 import com.vaadin.util.CurrentInstance;
 
 /**
@@ -76,11 +75,6 @@ import com.vaadin.util.CurrentInstance;
 public class UI implements Serializable, PollNotifier {
 
     public static final String POLL_DOM_EVENT_NAME = "ui-poll";
-
-    /**
-     * The Vaadin session to which this UI belongs.
-     */
-    private volatile VaadinSession session;
 
     /**
      * The id of this UI, used to find the server side instance of the UI form
@@ -116,11 +110,11 @@ public class UI implements Serializable, PollNotifier {
     }
 
     /**
-     * Gets the application object to which the component is attached.
+     * Gets the VaadinSession to which this UI is attached.
      *
      * <p>
-     * The method will return {@code null} if the component is not currently
-     * attached to an application.
+     * The method will return {@code null} if the UI is not currently attached
+     * to a VaadinSession.
      * </p>
      *
      * <p>
@@ -136,62 +130,7 @@ public class UI implements Serializable, PollNotifier {
      * @see #attach()
      */
     public VaadinSession getSession() {
-        return session;
-    }
-
-    /**
-     * Sets the session to which this UI is assigned.
-     * <p>
-     * This method is for internal use by the framework. To explicitly close a
-     * UI, see {@link #close()}.
-     * </p>
-     *
-     * @param session
-     *            the session to set
-     *
-     * @throws IllegalStateException
-     *             if the session has already been set
-     *
-     * @see #getSession()
-     */
-    public void setSession(VaadinSession session) {
-        if (session == null && this.session == null) {
-            throw new IllegalStateException(
-                    "Session should never be set to null when UI.session is already null");
-        } else if (session != null && this.session != null) {
-            throw new IllegalStateException(
-                    "Session has already been set. Old session: "
-                            + getSessionDetails(this.session)
-                            + ". New session: " + getSessionDetails(session)
-                            + ".");
-        } else {
-            if (session == null) {
-                try {
-                    detach();
-                } catch (Exception e) {
-                    getLogger().log(Level.WARNING,
-                            "Error while detaching UI from session", e);
-                }
-                // Disable push when the UI is detached. Otherwise the
-                // push connection and possibly VaadinSession will live on.
-                getPushConfiguration().setPushMode(PushMode.DISABLED);
-                getInternals().setPushConnection(null);
-            }
-            this.session = session;
-        }
-
-        if (session != null) {
-            attach();
-        }
-    }
-
-    private static String getSessionDetails(VaadinSession session) {
-        if (session == null) {
-            return null;
-        } else {
-            return session.toString() + " for "
-                    + session.getService().getServiceName();
-        }
+        return internals.getSession();
     }
 
     /**
@@ -334,8 +273,8 @@ public class UI implements Serializable, PollNotifier {
             // the UI is detached and cleaned up.
 
             // Can't use UI.push() directly since it checks for a valid session
-            if (session != null) {
-                session.getService().runPendingAccessTasks(session);
+            if (getSession() != null) {
+                getSession().getService().runPendingAccessTasks(getSession());
             }
             pushConnection.push();
         }
